requirements.txt
python-bitcoinrpc==1.0
mysql-connector-python==8.1.0
requests==2.31.0